fix(individual changelog): wrap existing article in container, navbar and footer
